Extract load_descriptor_from_textproto to common protobuf_utils.hpp
