+++ tools/scaleout/CMakeLists.txt
@@ -41,6 +41,7 @@ target_sources(
             board/board.hpp
             cabling_generator/cabling_generator.hpp
             cabling_generator/descriptor_merger.hpp
+            cabling_generator/protobuf_utils.hpp
             connector/connector.hpp
             factory_system_descriptor/utils.hpp
             node/node_types.hpp
