+++ tools/scaleout/cabling_generator/descriptor_merger.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "descriptor_merger.hpp"
+#include "protobuf_utils.hpp"
 
 #include <algorithm>
 #include <filesystem>
@@ -20,22 +21,6 @@
 namespace tt::scaleout_tools {
 
 namespace {
-
-template <typename Descriptor>
-Descriptor load_descriptor_from_textproto(const std::string& file_path) {
-    std::ifstream file(file_path);
-    if (!file.is_open()) {
-        throw std::runtime_error("Failed to open file: " + file_path);
-    }
-
-    const std::string file_content((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
-
-    Descriptor descriptor;
-    if (!google::protobuf::TextFormat::ParseFromString(file_content, &descriptor)) {
-        throw std::runtime_error("Failed to parse textproto file: " + file_path);
-    }
-    return descriptor;
-}
 
 }  // namespace
